DISCOVERYACCESS-7721 - move simplecov code to cucumber env.rb
