+++ features/support/env.rb
@@ -1,3 +1,19 @@
+if ( ENV['COVERAGE'] == 'on' )
+  require 'simplecov'
+  require 'simplecov-rcov'
+  class SimpleCov::Formatter::MergedFormatter
+    def format(result)
+       SimpleCov::Formatter::HTMLFormatter.new.format(result)
+       SimpleCov::Formatter::RcovFormatter.new.format(result)
+    end
+  end
+  SimpleCov.formatter = SimpleCov::Formatter::MergedFormatter
+  SimpleCov.start 'rails' do
+    add_filter "/vendor/"
+  end
+  puts "required simplecov (features/support/env.rb)"
+end
+
 require 'rubygems'
 if false
 require 'spork'
