+++ spec/spec_helper.rb
@@ -1,23 +1,3 @@
-if ( ENV['COVERAGE'] == 'on' )
-  require 'simplecov'
-  require 'simplecov-rcov'
-  class SimpleCov::Formatter::MergedFormatter
-    def format(result)
-       SimpleCov::Formatter::HTMLFormatter.new.format(result)
-       SimpleCov::Formatter::RcovFormatter.new.format(result)
-    end
-  end
-  SimpleCov.formatter = SimpleCov::Formatter::MergedFormatter
-  SimpleCov.start 'rails' do
-    add_filter "/vendor/"
-  end
-  puts "required simplecov (spec/spec_helper.rb)"
-else
-  require 'simplecov'
-  require 'simplecov-rcov'
-  SimpleCov.formatter = SimpleCov::Formatter::RcovFormatter
-end
-
 require 'rubygems'
 
 #require 'spork'
